mcp_server/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
beautifulsoup4===4.8.0
bs4
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
exceptiongroup==1.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
mcp==1.9.4
pydantic==2.11.7
pydantic-settings==2.9.1
pydantic_core==2.33.2
python-dotenv==1.1.0
python-multipart==0.0.20
requests==2.32.4
sniffio==1.3.1
soupsieve==2.7
sse-starlette==2.3.6
starlette==0.47.0
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.4.0
uvicorn==0.34.3
